modal confirm

+++ src/app/main/proposal/list/proposalList.component.ts
@@ -1,3 +1,4 @@
+import { ModalComfirmComponent } from './../../../modals/modal-comfirm/modal-comfirm.component';
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { FuseTranslationLoaderService } from '@fuse/services/translation-loader.service';
 import { fuseAnimations } from '@fuse/animations';
@@ -8,6 +9,7 @@ import { Router, ActivatedRoute } from '@angular/router';
 import { ColumnMode } from '@swimlane/ngx-datatable';
 import { ProposalService } from '../services/proposal.service';
 import { NgxSpinnerService } from 'ngx-spinner';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-proposal-list',
@@ -27,7 +29,8 @@ export class ProposalListComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private propService: ProposalService,
-    private spinner: NgxSpinnerService
+    private spinner: NgxSpinnerService,
+    public dialog: MatDialog
   ) {
     this._fuseTranslationLoaderService.loadTranslations(english, thai);
   }
@@ -37,27 +40,37 @@ export class ProposalListComponent implements OnInit {
     this.rows = this.route.snapshot.data.items.data;
   }
 
-  
+
 
   clickAdd() {
     this.router.navigateByUrl("/proposal/proposalForm/new");
   }
 
-  editProposal(ev){
-    switch (ev.type){
+  editProposal(ev) {
+    switch (ev.type) {
       case "edit":
         this.spinner.show();
         this.router.navigateByUrl("/proposal/proposalForm/" + ev.data._id);
         break;
       case "delete":
-        this.propService.deleteProposalData(ev.data).then((res)=>{
-          this.propService.getProposalDataList().subscribe((res:any)=>{
-            this.rows = res.data;
-          })
-        })
+        const dialogRef = this.dialog.open(ModalComfirmComponent, {
+          width: '250px',
+          data: { title: "ยืนยันการลบ", massage: "กรุณาตรวจสอบอีกรอบ" }
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+          // console.log(result);
+          if (result === "comfirm") {
+            this.propService.deleteProposalData(ev.data).then((res) => {
+              this.propService.getProposalDataList().subscribe((res: any) => {
+                this.rows = res.data;
+              })
+            })
+          }
+        });
         break;
     }
-    
+
   }
 
   updateFilter(event) {
@@ -65,7 +78,7 @@ export class ProposalListComponent implements OnInit {
     // const val = event.target.value.toLowerCase();
 
     // filter our data
-    
+
   }
 
 }
